requirements.txt
numpy
scipy
torch
pytorch-ignite
scikit-learn